+++ CHANGELOG.md
@@ -22,6 +22,11 @@ Some examples, more below in the actual changelog (newer entries are more likely
 * (internal) generic client: add hook FilterRequestURLHook (#123, @marioreggiori)
 
 -->
+
+### Added
+
+* api/mock/mock_api_implementation: panic when trying to fake `ResourceWithTag` (#419, @drpsychick)
+
 ## [0.7.6] -- 2024-11-04
 
 ### Added